+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Changelog
 ---------
 
+* Change all "NPY_IN_ARRAY" to "NPY_ARRAY_IN_ARRAY" *Sep 18 2025* to adapt numpy>2.3.3
+
 * **Pyseistr 0.0.2** *July 20 2022*
     * Release of pyseistr in aaspip